Ensure we always write to output stream 0
If an input has multiple streams, then the packet's stream index for the
video stream we are reading may not match our output's stream index.
This can lead to errors such as:

[mp4 @ 0x5631c5836000] Invalid packet stream index: 1

Update the packet's stream index, prior to writing it, to match what it
is in our output.

I also update a comment (around line 294) to clarify that while we're
skipping a packet, it's not because it's not video.